requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
python-dotenv
pydantic
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
Pillow
